ct-delivery-assistant-poc/requirements.txt
fastapi>=0.118.0
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
httpx==0.28.1
python-dotenv==1.0.1
itsdangerous==2.2.0
redis>=5.0.0
openai>=1.0.0
starlette>=0.49.1
